Move MockClient to source, update contract transactions to use ContractLikeId
