+++ __tests__/contract/ContractDeleteTransaction.test.ts
@@ -1,28 +1,5 @@
-import {BaseClient, unaryCall} from "../src/BaseClient";
-import {grpc} from "@improbable-eng/grpc-web";
-import {Ed25519PrivateKey} from "../src/Keys";
-import ProtobufMessage = grpc.ProtobufMessage;
-import {ContractDeleteTransaction} from "../src/contract/ContractDeleteTransaction";
-
-const privateKey = Ed25519PrivateKey.fromString('302e020100300506032b657004220420db484b828e64b2d8f12ce3c0a0e93a0b8cce7af1bb8f39c97732394482538e10');
-
-class MockClient extends BaseClient {
-    public constructor() {
-        super(
-            { 'nonexistent-testnet': { shard: 0, realm: 0, account: 3 } },
-            {
-                account: { shard: 0, realm: 0, account: 2 },
-                privateKey
-            }
-        );
-    }
-
-    public [unaryCall]<Rq extends ProtobufMessage, Rs extends ProtobufMessage>(): Promise<Rs> {
-        throw new Error('should not be called');
-    }
-}
-
-const mockClient = new MockClient();
+import {ContractDeleteTransaction} from "../../src/contract/ContractDeleteTransaction";
+import {mockClient, privateKey} from "../MockClient";
 
 describe("ContractDeleteTransaction", () => {
     it('serializes and deserializes correctly; ContractDeleteTransaction', () => {
@@ -37,7 +14,6 @@ describe("ContractDeleteTransaction", () => {
             .build()
             .sign(privateKey);
 
-        // const tx = new TextDecoder("ascii").decode(transaction.toProto().toS);
         const tx = transaction.toProto().toObject();
         expect(tx).toStrictEqual({
             body: undefined,
