+++ __tests__/contract/ContractExecuteTransaction.test.ts
@@ -1,28 +1,5 @@
-import {BaseClient, unaryCall} from "../src/BaseClient";
-import {grpc} from "@improbable-eng/grpc-web";
-import {Ed25519PrivateKey} from "../src/Keys";
-import ProtobufMessage = grpc.ProtobufMessage;
-import {ContractExecuteTransaction} from "../src/contract/ContractExecuteTransaction";
-
-const privateKey = Ed25519PrivateKey.fromString('302e020100300506032b657004220420db484b828e64b2d8f12ce3c0a0e93a0b8cce7af1bb8f39c97732394482538e10');
-
-class MockClient extends BaseClient {
-    public constructor() {
-        super(
-            { 'nonexistent-testnet': { shard: 0, realm: 0, account: 3 } },
-            {
-                account: { shard: 0, realm: 0, account: 2 },
-                privateKey
-            }
-        );
-    }
-
-    public [unaryCall]<Rq extends ProtobufMessage, Rs extends ProtobufMessage>(): Promise<Rs> {
-        throw new Error('should not be called');
-    }
-}
-
-const mockClient = new MockClient();
+import {ContractExecuteTransaction} from "../../exports";
+import {mockClient, privateKey} from "../MockClient";
 
 describe("ContractExecuteTransaction", () => {
     it('serializes and deserializes correctly; ContractExecuteTransaction', () => {
@@ -40,7 +17,6 @@ describe("ContractExecuteTransaction", () => {
             .build()
             .sign(privateKey);
 
-        // const tx = new TextDecoder("ascii").decode(transaction.toProto().toS);
         const tx = transaction.toProto().toObject();
         expect(tx).toStrictEqual({
             body: undefined,
