+++ jest.config.js
@@ -2,6 +2,7 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   verbose: true,
+  setupFiles: ["./__tests__/MockClient.ts"],
   // only test Typescript files
-  testMatch: ['**/__tests__/**/*.ts']
+  testMatch: ['**/__tests__/**/*.test.ts']
 };
